requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
chromadb
pypdf
supabase
langchain
langchain-google-genai
google-genai